Use hex to represent EUIDs

+++ radixdlt-java/src/main/java/com/radixdlt/client/assets/Asset.java
@@ -1,7 +1,6 @@
 package com.radixdlt.client.assets;
 
 import com.radixdlt.client.core.address.EUID;
-import java.math.BigInteger;
 import java.util.Objects;
 
 public class Asset {
@@ -9,8 +8,8 @@ public class Asset {
 	/**
 	 * Radix Token asset. TODO: Read from universe file. Hardcode for now.
 	 */
-	public static final Asset XRD = new Asset("TEST", 100000, new EUID(BigInteger.valueOf("TEST".hashCode())));
-	public static final Asset POW = new Asset("POW", 1, new EUID(BigInteger.valueOf(79416)));
+	public static final Asset XRD = new Asset("TEST", 100000, new EUID("TEST".hashCode()));
+	public static final Asset POW = new Asset("POW", 1, new EUID(79416));
 
 	private final String iso;
 	private final int subUnits;
